Organized file paths, created build, example and cleaning scripts
