+++ build.sh
@@ -1,0 +1,14 @@
+#!/bin/bash
+BUILD_DIR="build"
+
+# Create the build directory if it doesn't exist
+if [ ! -d "$BUILD_DIR" ]; then
+    mkdir "$BUILD_DIR"
+fi
+
+# Navigate to the build directory
+cd "$BUILD_DIR"
+
+# Run CMake and build
+cmake ..
+cmake --build .
