+++ clean.sh
@@ -1,0 +1,11 @@
+#!/bin/bash
+BUILD_DIR="build"
+OUTPUT_DIR="output"
+
+if [ -d "$BUILD_DIR" ]; then
+    rm -rf "$BUILD_DIR"
+fi
+
+if [ -d "$OUTPUT_DIR" ]; then
+    rm -rf "$OUTPUT_DIR"
+fi
