+++ example.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+chmod +x build/bin/example
+mkdir output
+./build/bin/example
